change cutoff for all expressions download to -0.1. This will include genes with no expression in any assayGroup.
This is in response to a user query. They are trying to correlate genes between different baseline experiments and it would be helpful to see the studied genes with no expression.

Note, for the web interface it is not possible to specify a cutoff of less than 0, so the user will never see genes without expression in the heatmap, and instead with get a "No expression found" message.