Task 156367994 (#72)
* Add the gene ID retrieved from the expression matrix as-is to identifier_search
Since an ID might not be in the bioentities index (either because it’s been retired from Ensembl or in a scaffold/contig we haven’t retrieved from Biomart) we can’t rely on either ensgene or any other property value being present, making it unfindable.
Note: for this feature to be effective we need to reindex the analytics collection

* A hack to search gene IDs not present in the bioentities collection. Revert after a full analytics index.

* Now that SolrInputDocumentInputStream is using getOrDefault we can return the same bioentity properties for all types of data points without fearing a NPE

* Test what’s really relevant

* We now search in the lowercase normalized field bioentity_identifier_search

* Remove unused implementation of mapToLinkText without isSpecies parameter and fix test

* Ignore a bad test

* Remove unnecessary stubbings

* Revert after reverting 41a789167d1731783bac95595454ac2d708fa9a3

* Use an Oryza sativa gene ID that’s actually in the analytics collection

* Add zk.host property to fix app startup

* Build ZooKeeper URL as a bean in applicationContext.xml, which gets fitlered with properties from pom.xml and VM properties

* Test DAS endpoint sensibly

* If a gene ID is in our experiments but not in our annotations display a fallback page
If a gene ID isn’t in the bioentity collection (i.e. no properties are retrieved) check if it’s in the analytics collection; if that’s the case add a link to Ensembl/EG/WBPS, we can’t really say much more than that

* Produce the right Ensembl/EG/WBPS link
The gene/transcript/protein/family link depends on the same prefix URL as the primary genome browser of each species, so we take advantage of that by adding some smarts to BioEntityCardProperties: the logic of the which link template to return is self-contained in each property using a lambda that produces a String. A good example of “functions as data”.

* Only return an empty string if the link depends on the species

* OLS links should use HTTPS

* Only species-dependent links should be empty if species is unknown

* Only species-dependent links should be empty if species is unknown (reprise)