Feature #158116018 - Search genes across species by symbols (#121)
* Added method to GeneSearchServiceDao to retrieve all gene IDs associated with a property value.

* Add logging messages to know what’s going on

* Extend GeneSearchService to support multiple gene IDs
The signature of the method has changed since the return value needs to be associated to each of the gene IDs. The parallelization strategy used is the most performant one after comparing a single SQL statement using IN, a parallel stream borrowing threads from the common ForkJoinPool and using CompletableFutures.

* Increase test coverage

* Bugfix: the StreamContext set in the raw tuple stream was a new one, and the instance created with a SolrClientCache was left unused

* Extend SolrQueryBuilder to specify the sort parameter in queries

* Add search streaming expression support to our SolrCloud mini-framework

* Make tests parameterized to have better feedback

* Extend GeneSearchService to support multiple gene IDs
The signature of the method has changed since the return value needs to be associated to each of the gene IDs. The parallelization strategy used is the most performant one after comparing a single SQL statement using IN, a parallel stream borrowing threads from the common ForkJoinPool and using CompletableFutures.

* Increase test coverage

* Generic create method in SolrCloudCollectionProxyFactory and refactor subclasses of CollectionProxy in one package

* Fix: point G2E collection proxy at the right collection alias and add basic tests

* Add test to remind client users of SearchStreamBuilder of the SolrQueryBuilder requirements for the query to work properly

* Remove generic types from decorator stream expressions
It doesn’t make sense for these types to be associated to a collection proxy, since the source of the TupleStream may come from different collections (e.g. inner join or intersect) or fields gone through a select and their names have been replaced by some arbitrary string.

* Add some documentation to the root package of the SolrCloud framework

* Guarantee that property name keys will be in lower case

* Fix after cherry picking from https://github.com/ebi-gene-expression-group/atlas/tree/feature/157048944-integrate-bioentity-information_2

* Clean cruft from test (we’re testing static methods)

* Add Gene ID search DAO
It takes some query parameters and returns a set of gene ID matches.

* Add a GeneQuery object that represents a query made to SCEA

* Reword comment about the general purpose of the class; explain why so little is logged

* Add a search service that provides a set of matching gene IDs for a GeneQuery object

* Fix bad references after rebase due to package renaming

* Two more rebase issues

* Set log levels for tests

* Set same logging levels in modules gxa/sc as in base

* Set Mockito strictness to lenient to allow more than one stubbing
I have no idea how this test passed before...

* Fix more test errors after rebasing

* When no cell IDs are passed return empty
Fixes test in http://bar:8085/browse/ATLAS-SCXACIDEVTASKS8-6

* Make SpeciesUtils a bit more useful by injecting real data from the species.json file

* Rename and repackage SuggesterService to SuggesterDao and deprecate current API
The wrongly-named service was interfacing directly with Solr and exposing implementation details to controllers

* Add a SuggesterService that invokes the SuggesterDao; add tests
Establish a boundary between the service the application will access and the DAO which connects to Solr for suggestions

* Move “special” sets of BioentityPropertyName to BioentityCollectionsProxy

* Add a new field used in SCEA to establish some precedence rules for suggested properties in the main search page

* Add class that transforms a set of suggestions to be displayed by React Select

* Manually return an empty stream if the query is less than three characters long

* Serialize suggestions to JSON in the server, so that the client needs to do as little as possible

* Add endpoint to offer gene ID suggestions tailored to scxa-gene-search-form

* Update Solr configsets

* Be a bit more liberal in the return type
Unfortunately Stream<Map> is not a superclass of Stream<ImmutableMap>, sorry Mr Postel

* Add mappings for a two-step, POST-to-GET, search controller
The POST mapping is the one called by scxa-gene-search-form

* Update test after bioentities collection rebuild

* Add which species species-specific IDs cover in the description

* Comment what is returned by each method (if only Java had type aliases...)

* Since this is a unit test, build the species without the SpeciesUtils class

* Add support to cross-species symbol search and species-specific non-Ensembl IDs
First we find matching gene IDs (or return with an empty list of results stating a reason of why nothing was found) and then we produce the usual list of experiment results. Heavily refactored to take away all the transformation to JSON: since we’re producing plain maps and lists, Gson::toJson is all we need.

* Fix capitalization of marker genes label

* Add attributes in model to populate form with the original search query in search results

* Add wrapper package for gene search form

* Include styles to properly show React-Select

* Show gene search form package instead of Atlas Autocomplete

* Dev bundles

* Make property_name_id_weight the same type as in the schema

* Remove stale imports

* Return an empty map if cells have no facets
E-MTAB-4888 breaks the assumption that all cells will have some facets and it’ll probably won’t happen, but we won’t show an error

* Added missing annotation for failing test class.